requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2
django-allauth==0.45.0
django-crispy-forms==1.12.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
Paperboy==1.0.1
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.0.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
virtualenv==20.6.0
virtualenv-clone==0.5.5
whitenoise==5.3.0
